Refactor : 홈 페이지 관련 컴포넌트 디렉터리 구조 개편
